// apps/honami-system-web/src/api/model/agentResponse.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Shiron.HonamiSystem.Server | v1
 * OpenAPI spec version: 1.0.0
 */
import type { AgentPersonaResponse } from "./agentPersonaResponse";

export interface AgentResponse {
	id: string;
	name: string;
	/** @nullable */
	description: string | null;
	persona: null | AgentPersonaResponse;
	requiredTools: string[];
	suggestedTools: string[];
	createdAt: string;
	updatedAt: string;
}